+++ pyppeteer/connection.py
@@ -113,8 +113,8 @@ class Connection(EventEmitter):
 
     async def _on_close(self) -> None:
         if not self._recv_fut.done():
-            self._recv_fut.cancel()
-            await self.connection.close()
+            if hasattr(self, 'connection'):  # may not have connection
+                await self.connection.close()
         for cb in self._callbacks.values():
             cb.set_exception(NetworkError('connection closed'))
         self._callbacks.clear()