+++ CHANGES.md
@@ -1,11 +1,26 @@
 History
 =======
 
+## Version 0.0.11 (next version)
+
+Catch up puppeteer v0.12.0
+
+* Remove `ElementHandle.evaluate`
+* Remove `ElementHandle.attribute`
+* Deprecate `Page.plainText`
+* Deprecate `Page.injectFile`
+* Add `Page.querySelectorAllEval`
+* Add `Page.select` and `Page.type`
+* Add `ElementHandle.boudingBox` and `ElementHandle.screenshot`
+* Add `ElementHandle.focus`, `ElementHandle.type`, and `ElementHandle.press`
+* Add `getMetrics` method
+* Add `offlineMode`
+
 ## Version 0.0.10 (2018-02-27)
 
 * Enable to import `launch` from package root
 * Change `browser.close` to coroutine function
-* Chatch up puppeteer v0.11.0
+* Catch up puppeteer v0.11.0
 
 ### Version 0.0.9 (2017-09-09)